+++ Lib/ToolsDir/loader.js
@@ -24,6 +24,12 @@ R.r = function (name, parent) {
     }
     else {
         var parts = name.split("/");
+        if (parts.length >= 2) {
+            if (parts[0].charCodeAt(0) == 64) {
+                parts[0] = parts[0] + "/" + parts[1];
+                parts.splice(1, 1);
+            }
+        }
         if (parts.length <= 1) {
             p = R.map[name.toLowerCase()];
         }